refactor: redefine lib structure
