+++ bsp_epidemic_suppression_model/model_utilities/simplified_scenario.py
@@ -1,0 +1,13 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class SimplifiedScenario:
+    R0: float
+    ssapp: float
+    ssnoapp: float
+    scapp: float
+    tsapp: float
+    tsnoapp: float
+    papp: float
+    xi: float
