+++ bsp_epidemic_suppression_model/simplified_algorithm/simplified_algorithm.py
@@ -1,44 +1,13 @@
-from dataclasses import dataclass
-from bsp_epidemic_suppression_model.utilities.functions_utils import round2
-from bsp_epidemic_suppression_model.utilities.distributions import gamma_cdf
-from bsp_epidemic_suppression_model.utilities.epidemic_data import alpha, beta
+from bsp_epidemic_suppression_model.math_utilities.functions_utils import round2
 
+from bsp_epidemic_suppression_model.model_utilities.simplified_scenario import (
+    SimplifiedScenario,
+)
 
-def approximated_suppressed_R(
-    R0: float, FTs_infty: float, FTc_infty: float, xi: float, ts: float
-):
-    fraction_of_R0_before_isolation_for_symptoms = gamma_cdf(ts, alpha=alpha, beta=beta)
-    symptoms_rescaling = fraction_of_R0_before_isolation_for_symptoms + (
-        1 - fraction_of_R0_before_isolation_for_symptoms
-    ) * (1 - xi * FTs_infty)
-    contacts_rescaling = 1 - xi * FTc_infty
-
-    R = R0 * symptoms_rescaling * contacts_rescaling
-    return R
-
-
-def approximated_FTcapp(
-    scapp: float,
-    FTapp_tim1_infty: float,
-    FTnoapp_tim1_infty: float,
-    tildepapp_tim1: float,
-):
-    FTcapp_ti_infty = (
-        tildepapp_tim1 * FTapp_tim1_infty + (1 - tildepapp_tim1) * FTnoapp_tim1_infty
-    ) * scapp
-    return FTcapp_ti_infty
-
-
-@dataclass
-class SimplifiedScenario:
-    R0: float
-    ssapp: float
-    ssnoapp: float
-    scapp: float
-    tsapp: float
-    tsnoapp: float
-    papp: float
-    xi: float
+from bsp_epidemic_suppression_model.simplified_algorithm.approximate_model_blocks import (
+    approximated_suppressed_R,
+    approximated_FTcapp,
+)
 
 
 def simplified_time_evolution(
@@ -110,17 +79,3 @@ def simplified_time_evolution(
         )
 
         print(recap)
-
-
-if __name__ == "__main__":
-    simplified_scenario = SimplifiedScenario(
-        R0=1,
-        ssapp=0.7,
-        ssnoapp=0.2,
-        scapp=0.8,
-        tsapp=6.5,
-        tsnoapp=8.5,
-        papp=0.6,
-        xi=0.9,
-    )
-    simplified_time_evolution(simplified_scenario=simplified_scenario, n_iterations=5)
