+++ scripts/r0.py
@@ -1,6 +1,6 @@
-from bsp_epidemic_suppression_model.utilities.epidemic_data import beta0
-from bsp_epidemic_suppression_model.utilities.plotting_utils import plot_functions
-from bsp_epidemic_suppression_model.utilities.functions_utils import RealRange
+from bsp_epidemic_suppression_model.model_utilities.epidemic_data import beta0
+from bsp_epidemic_suppression_model.math_utilities.plotting_utils import plot_functions
+from bsp_epidemic_suppression_model.math_utilities.functions_utils import RealRange
 import scipy.integrate as integrate
 
 tau_max = 30
