+++ scripts/reproduction_number_time_evolution.py
@@ -1,12 +1,14 @@
-from bsp_epidemic_suppression_model.utilities.epidemic_data import (
+from bsp_epidemic_suppression_model.model_utilities.epidemic_data import (
     make_scenario_parameters_for_asymptomatics_symptomatics_model,
 )
-from bsp_epidemic_suppression_model.utilities.scenario import Scenario
-from bsp_epidemic_suppression_model.utilities.functions_utils import (
+from bsp_epidemic_suppression_model.model_utilities.scenario import Scenario
+from bsp_epidemic_suppression_model.math_utilities.functions_utils import (
     DeltaMeasure,
     RealRange,
 )
-from bsp_epidemic_suppression_model.utilities.plotting_utils import plot_time_evolution
+from bsp_epidemic_suppression_model.model_utilities.plot_results import (
+    plot_time_evolution,
+)
 from bsp_epidemic_suppression_model.algorithm.time_evolution_with_severity import (
     compute_time_evolution_with_severity,
 )
