+++ scripts/reproduction_number_time_evolution_span_parameters.py
@@ -1,8 +1,8 @@
-from bsp_epidemic_suppression_model.utilities.epidemic_data import (
+from bsp_epidemic_suppression_model.model_utilities.epidemic_data import (
     make_scenario_parameters_for_asymptomatics_symptomatics_model,
 )
-from bsp_epidemic_suppression_model.utilities.scenario import Scenario
-from bsp_epidemic_suppression_model.utilities.functions_utils import (
+from bsp_epidemic_suppression_model.model_utilities.scenario import Scenario
+from bsp_epidemic_suppression_model.math_utilities.functions_utils import (
     DeltaMeasure,
     RealRange,
 )
@@ -10,7 +10,7 @@ from bsp_epidemic_suppression_model.algorithm.model_blocks import effectiveness_
 from bsp_epidemic_suppression_model.algorithm.time_evolution_with_severity import (
     compute_time_evolution_with_severity,
 )
-from bsp_epidemic_suppression_model.utilities.functions_utils import round2
+from bsp_epidemic_suppression_model.math_utilities.functions_utils import round2
 
 import warnings
